backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-cloud-storage==2.0.0
google-auth==2.3.0
python-dotenv==0.19.1